# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@split-go.com
    name: Split Go API Support
    url: http://www.split-go.com/support
  description: 分帳系統 API 文檔 - 支援群組管理、交易記錄、結算功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Split Go API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: 獲取當前用戶加入的所有群組，包含創建者資訊
      produces:
      - application/json
      responses:
        "200":
          description: 群組列表
          schema:
            properties:
              data:
                items:
                  properties:
                    created_at:
                      type: string
                    created_by:
                      type: integer
                    creator:
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        username:
                          type: string
                      type: object
                    description:
                      type: string
                    id:
                      type: integer
                    name:
                      type: string
                    updated_at:
                      type: string
                  type: object
                type: array
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 獲取用戶群組列表
      tags:
      - 群組
    post:
      consumes:
      - application/json
      description: 創建一個新的群組，創建者自動成為群組管理員
      parameters:
      - description: 群組資料
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 群組創建成功
          schema:
            properties:
              data:
                properties:
                  created_at:
                    type: string
                  created_by:
                    type: integer
                  creator:
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      username:
                        type: string
                    type: object
                  description:
                    type: string
                  id:
                    type: integer
                  name:
                    type: string
                  updated_at:
                    type: string
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤或群組名稱為空
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 創建新群組
      tags:
      - 群組
  /groups/{id}:
    delete:
      description: 刪除群組及其所有相關資料，只有創建者可以執行此操作
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群組刪除成功
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 只有群組創建者可以刪除群組
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 刪除群組
      tags:
      - 群組
    get:
      description: 獲取群組的詳細信息，包含成員列表和當前用戶角色
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 群組詳細資訊
          schema:
            properties:
              data:
                properties:
                  current_user_role:
                    type: string
                  group:
                    properties:
                      created_at:
                        type: string
                      created_by:
                        type: integer
                      creator:
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          username:
                            type: string
                        type: object
                      description:
                        type: string
                      id:
                        type: integer
                      name:
                        type: string
                      updated_at:
                        type: string
                    type: object
                  members:
                    items:
                      properties:
                        avatar:
                          type: string
                        id:
                          type: integer
                        name:
                          type: string
                        username:
                          type: string
                      type: object
                    type: array
                type: object
              error:
                type: boolean
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 不是群組成員
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 獲取群組詳細資訊
      tags:
      - 群組
    put:
      consumes:
      - application/json
      description: 更新群組的名稱和描述，需要管理員權限
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新資料
        in: body
        name: request
        required: true
        schema:
          properties:
            description:
              type: string
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 群組更新成功
          schema:
            properties:
              data:
                properties:
                  created_at:
                    type: string
                  created_by:
                    type: integer
                  creator:
                    properties:
                      id:
                        type: integer
                      name:
                        type: string
                      username:
                        type: string
                    type: object
                  description:
                    type: string
                  id:
                    type: integer
                  name:
                    type: string
                  updated_at:
                    type: string
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤或群組名稱為空
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 沒有管理員權限
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新群組資訊
      tags:
      - 群組
...
  /groups/{id}/members:
    post:
      consumes:
      - application/json
      description: 向群組添加新成員，需要管理員權限
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      - description: 成員資料，role可選值：member, admin
        in: body
        name: request
        required: true
        schema:
          properties:
            role:
              type: string
            user_id:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 成員添加成功
          schema:
            properties:
              data:
                properties:
                  role:
                    type: string
                  user_id:
                    type: integer
                type: object
              error:
                type: boolean
              message:
                type: string
            type: object
        "400":
          description: 請求格式錯誤或用戶ID為空
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 沒有管理員權限
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組或用戶不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "409":
          description: 用戶已是群組成員
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 添加群組成員
      tags:
      - 群組
  /groups/{id}/members/{userId}:
    delete:
      description: 從群組中移除成員，需要管理員權限，不能移除創建者
      parameters:
      - description: 群組ID
        in: path
        name: id
        required: true
        type: integer
      - description: 要移除的用戶ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成員移除成功
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "401":
          description: 未授權
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "403":
          description: 沒有管理員權限或不能移除創建者
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "404":
          description: 群組或成員不存在
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            properties:
              error:
                type: boolean
              message:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 移除群組成員
      tags:
      - 群組
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 在 header 中加入 "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"